Replace keycodes with scancodes most of the places
In SDL2 the keycodes are not bounded by a small upper value anymore,
so keeping pressed state in an array based on key code is not feasible
anymore. Scancodes does however have an small upper value, so we can use
that to keep the pressed state in an array.

There are differences of how keycodes and scancodes behave, the former
being tied to keyboard layout, while the latter is fixed regardless of
layout. The controls for one player should be in vicinity of each other,
so given that it actually seems beneficial to change for scancode so that
the controls wouldn't be moved around on the keyboard if the keyboard
layout is changed.

For typing names, the keycodes are still being used, also because in that
case the keystate is not stored to an array, and it makes sense to respect
keyboard layout here.